Assignment 3 API Submission
I am submitting this for this week.  I was able to duplicate Dr. Jariwala's work.

I tried to do the same on the weather API spending almost the entire week.  I could not get it to function. The challenge was it requires you to submit the city name to provide the results.  I was able to create the city list but not tie them together to provide an output.

So I backed up with short time left and tried to add features to this work by adding Park Hours and Entry Fees.  However, the basic loading of the additional data gave me an error. So I clearly due not understand something with connecting to API's.

I do not consider this my work as I followed Dr. Jariwala's implementation, but my weather API work is not functional and thus not worthy to submit.

I did at least setup my own key if that counts.

Not great work on this assignment!

Shane